refactor(undo): extract undo engine to internal/undo
Move RunUndoOperation, PreflightUndoConflicts, UndoResult, UndoConflictReport,
UndoConflictItem and all helper functions from internal/server/undo_engine.go
to new package internal/undo/engine.go.

Key changes:
- Created internal/undo/engine.go with all undo logic
- Added OnFileMovedFunc callback parameter to RunUndoOperation for
  Deluge integration (passed as NotifyDelugeAfterUndo from server)
- Updated internal/server/undo_engine.go to wrapper functions that
  re-export the public API from internal/undo
- Comprehensive unit tests in internal/undo/engine_test.go covering
  file moves, conflicts, idempotency, and callback invocation

This extraction removes the server dependency from core undo logic,
enabling reuse in other contexts and improving separation of concerns.
